pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clipwright"
version = "0.1.0"
description = "Playwright-driven short-form how-to video pipeline."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "DarkMatter" }]
keywords = ["video", "playwright", "ffmpeg", "tts", "captions", "short-form"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "typer>=0.12",
    "playwright>=1.45",
    "pillow>=10.0",
    "requests>=2.31",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
clipwright = "clipwright.cli:app"

[project.urls]
Homepage = "https://github.com/DarkMatter/clipwright"
Issues = "https://github.com/DarkMatter/clipwright/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
clipwright = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
